- 64-bit cleanup; action->{p1,p2,p3}.number changed from int to long

+++ route_struct.h
@@ -94,7 +94,7 @@ struct action{
 	int p2_type;
 	int p3_type;
 	union {
-		int number;
+		long number;
 		char* string;
 		void* data;
 	}p1, p2, p3;
